add new file for replication_node_manager
