+++ cubrid/CMakeLists.txt
@@ -211,6 +211,7 @@ set (REPLICATION_HEADERS
   ${REPLICATION_DIR}/slave_control_channel.hpp
   ${REPLICATION_DIR}/master_control_channel.hpp
   ${REPLICATION_DIR}/replication_node.hpp
+  ${REPLICATION_DIR}/replication_node_manager.hpp
   ${REPLICATION_DIR}/replication_object.hpp
   ${REPLICATION_DIR}/replication_slave_node.hpp
   ${REPLICATION_DIR}/replication_stream_entry.hpp
@@ -275,6 +276,7 @@ set(REPLICATION_SOURCES
   ${REPLICATION_DIR}/replication_master_node.cpp
   ${REPLICATION_DIR}/replication_master_senders_manager.cpp
   ${REPLICATION_DIR}/replication_node.cpp
+  ${REPLICATION_DIR}/replication_node_manager.cpp
   ${REPLICATION_DIR}/replication_slave_node.cpp
   ${REPLICATION_DIR}/replication_stream_entry.cpp
   )
