+++ src/replication/replication_master_node.cpp
@@ -132,88 +132,4 @@ namespace cubreplication
   }
 
   std::mutex master_node::g_enable_active_mtx;
-
-  std::mutex commute_mtx;
-
-  std::string host_name;
-  replication_node_manager *g_instance;
-
-  // Start as slave
-  replication_node_manager::replication_node_manager (const char *name)
-  {
-    INT64 buffer_size = prm_get_bigint_value (PRM_ID_REPL_CONSUMER_BUFFER_SIZE);
-    int num_max_appenders = log_Gl.trantable.num_total_indices + 1;
-    m_stream = new cubstream::multi_thread_stream (buffer_size, num_max_appenders);
-    m_stream->set_name ("repl" + std::string (name));
-    m_stream->set_trigger_min_to_read_size (stream_entry::compute_header_size ());
-    m_stream->init (0);
-
-    std::string replication_path;
-    replication_node::get_replication_file_path (replication_path);
-    m_stream_file = new cubstream::stream_file (*m_stream, replication_path);
-
-    m_repl_node = new slave_node (name, m_stream, m_stream_file);
-  }
-
-  replication_node_manager::~replication_node_manager ()
-  {
-    // stream and stream file are interdependent, therefore first stop the stream
-    m_stream->set_stop ();
-
-    delete m_stream_file;
-    delete m_stream;
-    delete m_repl_node;
-  }
-
-  void replication_node_manager::init_hostname (const char *name)
-  {
-    host_name = name;
-  }
-
-  replication_node_manager *replication_node_manager::get_instance ()
-  {
-    if (g_instance == NULL)
-      {
-	g_instance = new replication_node_manager (host_name.c_str ());
-      }
-
-    return g_instance;
-  }
-
-  void replication_node_manager::finalize ()
-  {
-    delete g_instance;
-    g_instance = NULL;
-  }
-
-  void replication_node_manager::commute_to_master_state ()
-  {
-    _er_log_debug (ARG_FILE_LINE, "Commuted to master\n");
-    delete m_repl_node;
-    m_repl_node = new master_node (host_name.c_str (), m_stream, m_stream_file);
-  }
-
-  void replication_node_manager::commute_to_slave_state ()
-  {
-    delete m_repl_node;
-    m_repl_node = new slave_node (host_name.c_str (), m_stream, m_stream_file);
-  }
-
-  master_node *replication_node_manager::get_master_node ()
-  {
-    std::lock_guard<std::mutex> lg (commute_mtx);
-    if (dynamic_cast<master_node *> (m_repl_node) == nullptr)
-      {
-	commute_to_master_state ();
-      }
-    return (master_node *) m_repl_node;
-  }
-
-  slave_node *replication_node_manager::get_slave_node ()
-  {
-    std::lock_guard<std::mutex> lg (commute_mtx);
-    assert (dynamic_cast<slave_node *> (m_repl_node) != nullptr);
-    return (slave_node *) m_repl_node;
-  }
-
 } /* namespace cubreplication */
