+++ src/replication/replication_node_manager.hpp
@@ -18,42 +18,46 @@
  */
 
 /*
- * replication_master_node.hpp
+ * replication_node_manager.hpp
  */
 
-#ident "$Id$"
 
-#ifndef _REPLICATION_MASTER_NODE_HPP_
-#define _REPLICATION_MASTER_NODE_HPP_
+#ifndef _REPLICATION_NODE_MANAGER_HPP_
+#define _REPLICATION_NODE_MANAGER_HPP_
 
-#include "replication_node.hpp"
-#include "stream_file.hpp"
-#include "multi_thread_stream.hpp"
-
-#include <mutex>
+namespace cubstream
+{
+  class stream_file;
+  class multi_thread_stream;
+}
 
 namespace cubreplication
 {
-  class master_ctrl;
+  class master_node;
+  class replication_node;
   class slave_node;
 
-  class master_node : public replication_node
+  class replication_node_manager
   {
-    private:
-      static std::mutex g_enable_active_mtx;
-
     public:
-      master_node (const char *nam, cubstream::multi_thread_stream *stream, cubstream::stream_file *stream_file);
-      ~master_node ();
-      master_ctrl *m_control_channel_manager;
+      static void init_hostname (const char *name);
+      static replication_node_manager *get_instance ();
+      static void finalize ();
 
-      void init (const char *name);
-      void new_slave (int fd);
-      void add_ctrl_chn (int fd);
-      void final (void);
-      void enable_active (void);
-      void update_senders_min_position (const cubstream::stream_position &pos);
+      void commute_to_master_state ();
+      void commute_to_slave_state ();
+
+      master_node *get_master_node ();
+      slave_node *get_slave_node ();
+
+    private:
+      replication_node_manager ();
+      ~replication_node_manager ();
+
+      cubstream::stream_file *m_stream_file;
+      cubstream::multi_thread_stream *m_stream;
+      cubreplication::replication_node *m_repl_node;
   };
-} /* namespace cubreplication */
+}
 
-#endif /* _REPLICATION_MASTER_NODE_HPP_ */
+#endif
